Add `raceAndCancel` helper function for structured concurrency.
Motivation:

The core concept of Structured concurrency [1] requires that
"all spawned threads have completed before exit".
The classic implementation of `Promise.race` doesn't meet this
requirement, as the returned promise finishes sooner than any
of the other promises and the other promises are kept running
until completion. Therefore, the `Promise.race` violates the
basic principle of structured concurrency when used as a building
block for combining parallel tasks.

We propose a simple wrapper for the `Promise.race` that ensures
all remaining promises are automatically cancelled upon completion
of the first one. We name the wrapper `raceAndCancel`, however,
we are open to any other naming suggestions.

Example of use:

```
	const result = await raceAndCancel([
		someLongOperation(),
		CancellablePromise.delay(timeout)
	])
```

[1] https://en.wikipedia.org/wiki/Structured_concurrency